pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fusionclaw"
version = "0.1.0"
description = "Context fusion beats agent chat. Merge agent context windows instead of making them talk."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "FusionClaw Contributors"}]
keywords = ["agents", "llm", "multi-agent", "context-fusion", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic>=2.0",
    "openai>=1.0",
    "tiktoken>=0.7",
]

[project.optional-dependencies]
dev = [
    "python-dotenv>=1.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
benchmark = [
    "python-dotenv>=1.0",
]

[project.urls]
Repository = "https://github.com/fusionclaw/fusionclaw"

[tool.setuptools.packages.find]
include = ["fusionclaw*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
